% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_baseline_column.R
\name{add_baseline_column}
\alias{add_baseline_column}
\title{add_baseline_column}
\usage{
add_baseline_column(df, preferred_baseline = NULL)
}
\arguments{
\item{df}{km dataframe where baseline is to be added}

\item{preferred_baseline}{specifying a desired baseline}
}
\value{
returns a dataframe
}
\description{
given a data.frame with treatments and studies
return the same df but with extra column 'baseline'
that automatically selects some baseline for each study
you can choose your preferred 'global' baseline manually
}
