% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_connected}
\alias{check_connected}
\title{check_connected}
\usage{
check_connected(df)
}
\arguments{
\item{df}{the input dataframe used for carrying the network meta-analysis}
}
\value{
boolean indicating whether network is connected or not. A valid datafame would return TRUE.
}
\description{
A simple check to see if the network provided is connected or not.
Meta-Analysis should be carried only on connected networks. It is used directly in survnma as
a stopping mechanism for networks that are not valid.
}
