% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guyot.R
\name{guyot.method}
\alias{guyot.method}
\title{Go from KM curve data from published figures into individual-level data}
\usage{
guyot.method(
  x,
  y,
  t,
  r,
  km_write_path = NULL,
  ipd_write_path = NULL,
  tot.events = NA
)
}
\arguments{
\item{x}{X axis coordinates taken from the digitised Kaplan-Meier curve}

\item{y}{Y axis coordinates taken from the digitised Kaplan-Meier curve; same length as \code{x}}

\item{t}{times at which number at risk \code{r} is given}

\item{r}{number at risk; same length as \code{t}}

\item{tot.events}{total number of events reported (integer)}
}
\description{
This function is a simple adaptation of code given by Guyot et al.
}
\references{
Guyot, Patricia, AE Ades, Mario JNM Ouwens, and Nicky J. Welton.
“Enhanced Secondary Analysis of Survival Data: Reconstructing the
Data from Published Kaplan-Meier Survival Curves.”
BMC Medical Research Methodology 12, no. 1 (February 1, 2012): 9.
https://doi.org/10.1186/1471-2288-12-9.
}
