% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_plot.R
\name{hazard_plot}
\alias{hazard_plot}
\title{hazard_plot}
\usage{
hazard_plot(
  data = NULL,
  treatments = NULL,
  reference = NULL,
  timesteps = seq(0, 30),
  study_labs = NA,
  colors = NULL,
  xlab = "",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  title = "",
  interval = FALSE
)
}
\arguments{
\item{data}{a dataframe obtained directly through function \code{\link{prep_all_hazards}}
or a survnma object}

\item{treatments}{a list of treatments you want to plot}

\item{reference}{string representing reference treatment}

\item{timesteps}{Vector of evaluation times. Should be of length > 2}

\item{study_labs}{a collection of labels for the studies}

\item{colors}{a collection of colors for graphing}

\item{xlab}{a string to be used for labelling the x-axis}

\item{ylab}{a string to be used for labelling the y-axis}

\item{xlim}{specifying a limit for the x-axis}

\item{ylim}{specifying a limit for the y-axis}

\item{title}{title to be used for the hazard plot}

\item{interval}{boolean whether you want a credible interval on the curve}
}
\value{
a figure of required hazard-ratio plots
}
\description{
A function to be used with data* passed through prep_all_hazards. It plots the hazard-ratio curves
for indicated treatments. Alternatively, one can use a survnma object directly. The function will
then calculate the hazards internally. Execution is identical in either case.
}
