% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_table.R
\name{hazard_table}
\alias{hazard_table}
\title{hazard_table}
\usage{
hazard_table(survnma, treatments = NULL, reference, timesteps = seq(0,
  30), hazards = NULL, print = FALSE)
}
\arguments{
\item{survnma}{survnma object to be passed}

\item{treatments}{a list of desired treatments to be compared}

\item{reference}{string of treatment that acts as reference}

\item{timesteps}{Vector of evaluation times. Should be of length > 2}

\item{hazards}{precalculated hazards (using \code{\link{prep_all_hazards}})
to reduce computation time}

\item{print}{Boolean, if true prints results in console, if false stores dataframe}
}
\description{
internal function that generates hazard ratio table at each month
if you have pre-calculated hazard values you can pass
them directly to reduce computation time.
}
