% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_all_hazards.R
\name{prep_all_hazards}
\alias{prep_all_hazards}
\title{prep_all_hazards}
\usage{
prep_all_hazards(nmafit, treatments = NULL, reference, timesteps = seq(0, 30))
}
\arguments{
\item{nmafit}{a \code{survnma} object to be used for subsequent calculations}

\item{treatments}{a string or vector of treatments to be compared}

\item{reference}{a base treatment to be used for comparison when calculating
the hazard ratio}

\item{timesteps}{Vector of evaluation times. Should be of length > 2}
}
\value{
a dataframe with columns corresponding to
\code{c("curve", "time", "mean", "lci", "uci", "label")}
where curve indicates the family of distribution used for
survnma object and label indicating the corresponding
treatment considered. uci and lci indicate the upper and
lower bounds of the credible interval for the mean.
the dataframe is to be used for plotting via the \code{\link{hazard_plot}} function
}
\description{
A function that computes the hazard function and thus
hazard ratio for multiple treatments based on a study
of interest
}
