% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_all_survivals.R
\name{prep_all_survivals}
\alias{prep_all_survivals}
\title{prep_all_survivals}
\usage{
prep_all_survivals(nmafit, study, treatments = nmafit$treatment,
  timesteps = seq(0, 30))
}
\arguments{
\item{nmafit}{a survnma object to be used for subsequent calculations}

\item{study}{a string of the study of interest}

\item{treatments}{a string or vector of treatments to be compared}

\item{timesteps}{Vector of evaluation times. Should be of length > 2}
}
\value{
a dataframe with columns corresponding to
\code{c("curve", "time", "mean", "lci", "uci", "label")}
where curve indicates the family of distribution used for
survnma object and label indicating the corresponding
treatment considered. uci and lci indicate the upper and
lower bounds of the credible interval for the mean.
the dataframe is to be used for plotting via the \code{\link{survival_plot}} function
}
\description{
A function that computes the survival function for multiple treatments based on a study
of interest
}
