% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_survnma.R
\name{print.survnma}
\alias{print.survnma}
\title{Print out survnma object
nothing fancy for now, S3 method
to stop from printing all of survnma
object with its inputs etc. etc.}
\usage{
\method{print}{survnma}(x, ...)
}
\arguments{
\item{x}{Object of class \code{survnma}}

\item{...}{additional arguments to print function}
}
\description{
Print out survnma object
nothing fancy for now, S3 method
to stop from printing all of survnma
object with its inputs etc. etc.
}
