% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_km_folder.R
\name{read_km_folder}
\alias{read_km_folder}
\title{Read many digitized Kaplan-Meier curves}
\usage{
read_km_folder(path, preferred_baseline = NULL, filter = "KM.txt",
  split = "_", study = 1, outcome = 2, treatment = 3)
}
\arguments{
\item{path}{path to a folder containing txt files from digitizing; each KM file
should have 4 columns in the following order: time, number at risk,
number censored, number of events}

\item{preferred_baseline}{character; if you want to specify a preferred baseline treatment}

\item{filter}{character; only files containing \code{filter} in title will be used}

\item{split}{character; sign that separates study, outcome and treatment}

\item{study}{integer explaining naming convention, see below}

\item{outcome}{integer explaining naming convention, see below}

\item{treatment}{integer explaining naming convention, see below}
}
\value{
Data frame which can be used as input into NMA models using \code{\link{survnma}}.
}
\description{
The function reads a collection of files containing data from digitized Kaplan-Meier curves.
}
\details{
By default we assume that the files will follow naming convention such as
\code{study_outcome_trt_KM.txt}, e.g. \code{MOT_OS_Pazo_KM.txt}. If the naming
scheme is different, user can reorder at will or change the separation sign, e.g.
for \code{OS-Pazo-MOT-KM.txt} we would say
\code{split = "-", study = 3, outcome = 1, treatment = 2}.
}
