% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_curve.R
\name{survival_curve}
\alias{survival_curve}
\title{survival_curve}
\usage{
survival_curve(family, param.mtx, times, P = NULL, type = "interval")
}
\arguments{
\item{family}{character; any of the following distributional families:
\code{c("weibull", "gompertz", "exponential", "log-normal", "log-logistic", "fp1", "fp2")}}

\item{param.mtx}{is a \code{[n x m]} numeric matrix, where \code{m} is the number of parameters
\code{n} is number of samples from their underlying distribution}

\item{times}{vector of timepoints at which to calculate survival S(t)}

\item{P}{vector of powers, used only if \code{family == "fp1"} (uses \code{p[1]})
or \code{family == "fp2"} (uses \code{p[1]} and \code{p[2]})}

\item{type}{string: interval, mean or all}
}
\value{
a vector \code{c(q1, median, q3, mean)}
}
\description{
Survival curve calculation for parametric families
}
