% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_plot.R
\name{survival_plot}
\alias{survival_plot}
\title{survival_plot}
\usage{
survival_plot(
  data = NULL,
  study = NULL,
  treatments = NULL,
  timesteps = seq(0, 30),
  study_labs = NA,
  colors = NULL,
  xlab = NULL,
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  title = "",
  interval = TRUE
)
}
\arguments{
\item{data}{One of the two:
\itemize{
\item{a data-frame object (e.g. from \code{\link{prep_all_survivals}} function) containing columns
of \code{c("curve","time", "mean", "lci", "uci", "label")}}
\item{\code{survnma} object; this will lead to "behind-the-scenes"
survival curve calculations;
in that case you should/can also define \code{study},
\code{treatments}, \code{timesteps} (see \code{\link{prep_all_survivals}})}}}

\item{study}{a specified study to be used for carrying the analysis}

\item{treatments}{a selection of treatments to be compared; can be the whole set of available treatments}

\item{timesteps}{Vector of evaluation times. Should be of length > 2}

\item{study_labs}{labels for each study}

\item{colors}{vector containing color codes to be used for curves}

\item{xlab}{labels for x-axis}

\item{ylab}{labels for y-axis}

\item{xlim}{a limit for the x-axis}

\item{ylim}{a limit for the y-axis}

\item{title}{label for title}

\item{interval}{boolean for producing credible interval or not}
}
\value{
returns a graph of survival curves
}
\description{
Survival plotting function for survnma objects
}
