% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_table.R
\name{survival_table}
\alias{survival_table}
\title{survival_table}
\usage{
survival_table(survnma, study, treatments = NULL, timesteps = seq(0,
  30), survivals = NULL, print = FALSE)
}
\arguments{
\item{survnma}{survnma object to be passed}

\item{study}{study of interest}

\item{treatments}{a list of desired treatments to be compared}

\item{timesteps}{Vector of evaluation times. Should be of length > 2}

\item{survivals}{precalculated survivals (using prep_all_survivals)
to reduce computation time}

\item{print}{Boolean, if true prints results in console, if false stores dataframe}
}
\description{
internal function that evaluates survivals at each month
if you have pre-calculated survival values you can pass
them directly to reduce computation time.
}
