% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survnma.R
\name{survnma}
\alias{survnma}
\title{Survival network meta-analysis models}
\usage{
survnma(nma_df = NULL, model = "exponential", type = "fixed",
  prior = NULL, P = NULL, inits = "generate", n.chains = 3,
  n.iter = 5000, auto_restart = FALSE, connected_check = TRUE,
  warnings = TRUE, bugs.directory = "C:/WinBUGS14",
  min_time_change = 0, ...)
}
\arguments{
\item{nma_df}{a \code{data.frame} with columns \code{treatment}, \code{study}, \code{baseline} and \code{filepath};
each file in \code{filepath} should have 4 columns in the following order: time, number at risk,
number censored, number of events;
refer to \link{read_km_folder} for automatically creating inputs from folder of text files}

\item{model}{\code{string}; family of survival curves to use
Use one of: \code{weibull}, \code{gompertz}, \code{exponential}
\code{lognormal}, \code{loglogistic}, \code{fp1}, \code{fp2}.
The last two refer to fractional polynomials of order one and two.}

\item{type}{a \code{string} of characters indicated whether you are considering
"fixed" or "random" effects when modelling the data}

\item{prior}{a list of prior values, with elements \code{mean}, \code{prec2} and \code{R} to be passed to WinBUGS;
if \code{prior=NULL}, the defaults are used;
dimensionality of values must be same as of the model
(1 for exponential, 3 for fp2, 2 for the rest);
\code{mean} is mean treatment effect (TE), default is 0;
\code{prec2} is precision of TE, default is 0.001*Identity
\code{R} is scale matrix of Wishart, only used in random effects model, default is Identity;
if using informative priors, we recommend setting \code{inits=NULL} or generating by hand}

\item{P}{\code{numeric} or \code{vector} of numbers indicating the powers to
be used when considering fractional polynomials (can be \code{NULL})}

\item{inits}{if \code{inits="generate"}, initial values will be generated
automatically within R, using Normal(0, 1) for d1/mu1 and Normal(0, 0.1^2)
for d2/mu2 values;
if \code{inits=NULL}, they will be generated automatically by WinBUGS
and much more variable than when generated within R;
if list is supplied, it will be used in a standard WinBUGS way
(see \code{\link[R2WinBUGS]{bugs}})}

\item{n.chains}{number of chains for WinBUGS to use, default is 3}

\item{n.iter}{integer specifying how many interations to be ran}

\item{auto_restart}{Boolean to continue running model until convergence or
stop after 5 iterations}

\item{connected_check}{logical; if FALSE, does not check if network is connected}

\item{warnings}{logical; if FALSE, warning messages about lack of convergence are skipped}

\item{bugs.directory}{WinBUGS directory path; defaults to \code{C:/WinBUGS14}}

\item{min_time_change}{numeric; a threshold for merging data points where differences in time are
very small (see Details). We suggest a range of 0.01-0.05.
Default of 0 does not modify any data.}

\item{...}{additional parameters passed to \code{\link[R2WinBUGS]{bugs}}
(e.g. \code{inits}, \code{n.iter})}
}
\value{
a \code{survnma} class object
}
\description{
survnma provides interface for Bayesian NMA WinBUGS models
(specifically through \code{R2WinBUGS}). It generates samples
from the posterior distribution and interprets them in terms of survivals and hazards.
}
\details{
WinBUGS implementation is through \code{R2WinBUGS} package. Additional arguments can be passed via \code{...},
including \code{bugs.directory}. We use default of \code{C:/WinBUGS14/}.

WinBUGS models code follows publications of M.J. Ouwens and J.P. Jansen (see References).
Models can be viewed with \code{\link[survnma]{survnma_model}}.

Input data have to follow a particular format, typically obtained from \link{read_km_folder} but
\code{data.frame} of inputs can also be specified manually.

Both initial values and priors can be specified by hand. To understand how they are used,
it is recommended to inspect the model with \code{\link[survnma]{survnma_model}} first. The
default generation of initial values (\code{inits = "generate"}) is optimised toward default
vague priors. When using informative priors it is best to set \code{inits = NULL}.
}
\examples{
# We read data from some included data files:
nma_df <- data.frame(
             stringsAsFactors = FALSE,
             "treatment" = c("Suni", "Ifn", "Suni", "Pazo"),
             "study" = c("Study 1", "Study 1", "Study 2", "Study 2"),
             "baseline" = c("Suni", "Suni", "Suni", "Suni"),
             "filepath" = sapply(c("Mota_OS_Suni_KM.txt",
                                   "Mota_OS_Ifn_KM.txt",
                                   "Mot_OS_Suni_KM.txt",
                                   "Mot_OS_Pazo_KM.txt"), function(x)
              system.file("extdata", "narrow", x, package="survnma", mustWork=TRUE)))

# example of simple survnma models (fixed effects)
fit_wbl <- survnma(nma_df, "weibull", min_time_change = 0.05)
fit_fp2 <- survnma(nma_df, "fp2", P = c(-1,0), min_time_change = 0.05)

#manually setting a (very informative) prior
#dimension of prior is 2 as Weibull is 2-parameter dist.
fit_wbl <- survnma(nma_df, "weibull",
                   prior = list("mean" = c(1,1), "prec2" = .1*diag(2)),
                   min_time_change = 0.05)

#manually setting some WinBUGS options
\donttest{survnma(nma_df, "weibull", n.iter = 10000, debug = TRUE)}

}
\references{
(1) Ouwens, Mario J. N. M., Zoe Philips, and Jeroen P. Jansen.
“Network Meta-Analysis of Parametric Survival Curves.”
Research Synthesis Methods 1, no. 3–4 (July 2010): 258–71. https://doi.org/10.1002/jrsm.25.

(2) Jansen, Jeroen P. “Network Meta-Analysis of Survival Data with Fractional Polynomials.”
BMC Medical Research Methodology 11, no. 1 (May 6, 2011): 61. https://doi.org/10.1186/1471-2288-11-61.
}
\seealso{
\code{vignette(survivalnma_guide)} for overview,
for working with outputs \code{\link[survnma]{survival_plot}}
and \code{\link[survnma]{hazard_plot}}
}
\author{
Witold Wiecek, Savvas Pafitis
}
