% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survnma_model.R
\name{survnma_model}
\alias{survnma_model}
\title{survnma_model}
\usage{
survnma_model(model, type)
}
\arguments{
\item{model}{one of \code{fp1}, \code{fp2}, \code{weibull}, \code{exponential}, \code{loglogistic}, \code{lognormal}, \code{gompertz}}

\item{type}{\code{fixed} or \code{random}}
}
\description{
Show the WinBUGS model used by survnma in a separate window.
}
